Implemented server unavailable message as overridable component
